Make signature and usage of computeStateType independent of Op type.
The current design of the IteratorAnalysis consists of a TypeSwitch
where each op (or at least group of op with the same number of operands)
has some glue code and then calls an op-specific function
`computeStateType`. This not only makes the signature of that function
dependent on the number of operands the corresponding op has, it also
distributes op-specific code over that function *and the TypeSwitch*,
both of which is unnecessary.

This commit changes the signature of `computeStateType` to use an array
of upstream state types, which may have zero, one, or more elements (one
for each operand).